split into multiple files and introduce ABC
