+++ umbridge/__init__.py
@@ -1,2 +1,4 @@
 # init file for umbridge
-from .um import *
+from .model import Model
+from .server import serve_models
+from .client import HTTPModel, supported_models
